// src/utils/enums/errorMessageEnum.ts
enum ErrorMessageEnum {
  INVALID_REQUEST = "Invalid Request Created",
  RECORD_NOT_FOUND = "Record Not Found",
  INVALID_USER_ID = "Invalid User Id",
  INVALID_EMAIL_OR_CODE = "Invalid Email or Code!!",
  SET_YOUR_PASSWORD = "Please set you password first!!",
  INVALID_CREDENTIALS = " Invalid credentials"
}

export default ErrorMessageEnum;